FROMLIST: clk: rockchip: associate SCLK_MAC_PLL and disable reparenting on rk3036
The emac needs constant and very specific rate but the possible PLL-sources
are very limited, so we expect the PLL source to be set manually on per
board and don't want it to get changed in an automatic way later.
So add the necessary clock-id and disable reparenting on set_rate calls.

Change-Id: I7e0ac383a6e35c2295692a26b5a508440fba03e3

Signed-off-by: Heiko Stuebner <heiko@sntech.de>

Signed-off-by: Caesar Wang <wxt@rock-chips.com>

(am from https://patchwork.kernel.org/patch/8573381/)
Signed-off-by: Caesar Wang <wxt@rock-chips.com>